Added File Handler to steamTradingHelprV2
Handles Local JSON file which stores link of collection and cases of csgostash. Creates a JSON file if not available. If JSON file not readable or changed, overwrites with new links from CSGOSTASH Scrapping. Hands the link to dictionary (csgostash_link_database) if everything goes correct in saves JSON file.